# butler/public/docs/openapi.yaml
openapi: 3.1.0
info:
  title: butler.works API (리버스 엔지니어링) + butler.view 로컬 API
  version: "1.0.0"
  description: |
    `www.butler.works` 의 내부 API(`api.butler.works`)를 HAR 캡처로 리버스 엔지니어링해
    정리한 명세 + 본 프로젝트(butler.view)가 제공하는 로컬 DB 기반 API.

    ## 업스트림(butler.works) 핵심 사실
    - 인증: 공개 엔드포인트는 토큰 불필요. CORS `Access-Control-Allow-Origin: *`.
    - 레이트리밋(비로그인): `x-ratelimit-limit-user: 100` / `x-ratelimit-reset-user: 60`
      → 사용자(IP)당 60초에 100요청.
    - 기업 식별자 `corpCode` = **DART 고유번호 8자리** (예: `00126380`=삼성전자).
      종목코드 `stockCode` 는 6자리(예: `005930`).
    - **데이터 게이팅**: 목표주가/컨센서스 리포트(증권사별)는 비로그인도 최신까지 공개.
      재무 시계열(매출/영업이익/순이익, PER/PBR 과거값)은 **비로그인 시 ~2023년까지만**
      내려오고, 로그인(구독) 세션에서만 최신 분기(2024~)까지 제공된다. 응답의
      `isLegacyMasking: true` 가 마스킹 신호.

    ## 데이터 출처(중요)
    butler 는 외부 컨센서스 피드를 그대로 중계하지 않고 증권사 리포트를 자체 수집·정규화한다.
    리포트 1건 = 증권사 1곳의 한 시점 목표주가. 월별 min/avg/max 는 butler 자체 집계.
    본 프로젝트의 모든 적재 데이터는 `source = "butler"` 로 출처를 표기한다.
servers:
  - url: https://api.butler.works
    description: butler.works 업스트림(공개)
  - url: http://localhost:3939/api/proxy
    description: butler.view 투명 프록시 (→ api.butler.works/api)
  - url: http://localhost:3939/api
    description: butler.view 로컬 DB API (정규화 SQLite)
...
paths:
...
  /sectors/{code}:
    get:
      tags: [sector]
      summary: 섹터 상세 (집계 + 90일 목표가 상향/하향 모멘텀 + 구성 기업)
      servers: [{ url: "http://localhost:3939/api" }]
      parameters:
        - { name: code, in: path, required: true, schema: { type: string }, example: semiconductor }
        - { name: sort, in: query, schema: { type: string, enum: [market_cap, target_return_rate] } }
      responses: { "200": { description: OK }, "404": { description: 섹터 없음 } }

  # ===================== 로컬: 인증/관심목록/알림 =====================
  /auth/signup:
    post:
      tags: [auth]
      summary: 회원가입 (이메일+비밀번호) → 세션 쿠키
      servers: [{ url: "http://localhost:3939/api" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
      responses: { "200": { description: OK, headers: { Set-Cookie: { schema: { type: string }, description: butler_session (httpOnly) } } }, "400": { description: 검증 실패 } }
...
  /watchlist:
    get:
      tags: [auth]
      summary: 내 관심목록
      servers: [{ url: "http://localhost:3939/api" }]
      responses: { "200": { description: OK }, "401": { description: 로그인 필요 } }
    post:
      tags: [auth]
      summary: 관심 기업 추가
      servers: [{ url: "http://localhost:3939/api" }]
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [corpCode], properties: { corpCode: { type: string } } }
      responses: { "200": { description: OK }, "401": { description: 로그인 필요 } }
    delete:
      tags: [auth]
      summary: 관심 기업 제거
      servers: [{ url: "http://localhost:3939/api" }]
      parameters: [{ name: corpCode, in: query, required: true, schema: { type: string } }]
      responses: { "200": { description: OK }, "401": { description: 로그인 필요 } }
...